requirements.txt
anyio==3.7.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==2.0.3
